requirements.txt
aiomysql>=0.2.0
certifi>=2024.8.30
cffi>=1.17.1
charset-normalizer>=3.4.0
cryptography>=43.0.3
dnspython>=2.7.0
idna>=3.10
motor>=3.6.0
psycopg2-binary>=2.9.10
pymongo>=4.9.2
requests>=2.32.3
urllib3>=2.2.3
